--- calculator/addition/BUILD
@@ -1,9 +1,0 @@
-load("@rules_cc//cc:defs.bzl","cc_binary","cc_library")
-
-cc_library(
-
-    name = "addition",
-    srcs= ["addition.cpp"],
-    hdrs = ["addition.h"],
-    visibility = ["//visibility:public"],
-)
