--- calculator/addition/addition.cpp
@@ -1,6 +1,0 @@
-#include "addition.h"
-
-float addition(float a, float b)
-{
-    return a+b;
-}
