--- calculator/addition/addition.h
@@ -1,6 +1,0 @@
-#ifndef CALCULATOR_ADDITION_H
-#define CALCULATOR_ADDITION_H
-
-float addition(float a,float b);
-
-#endif // ADDITION_H
