--- calculator/division/BUILD
@@ -1,9 +1,0 @@
-load("@rules_cc//cc:defs.bzl","cc_binary","cc_library")
-
-cc_library(
-
-    name = "division",
-    srcs= ["division.cpp"],
-    hdrs = ["division.h"],
-    visibility = ["//visibility:public"],
-)
