--- calculator/division/division.cpp
@@ -1,6 +1,0 @@
-#include "division.h"
-
-float division(float a, float b)
-{
-    return a/b;
-}
