--- calculator/division/division.h
@@ -1,6 +1,0 @@
-#ifndef CALCULATOR_DIVISION_H
-#define CALCULATOR_DIVISION_H
-
-float division(float a, float b);
-
-#endif // DIVISION_H
