+++ calculator/functionalities.cpp
@@ -1,0 +1,26 @@
+#include <iostream>
+#include "functionalities.h"
+
+
+float addition(float a, float b)
+{
+    return a+b;
+}
+
+float division(float a, float b)
+{
+    return a/b;
+}
+
+float multiplication(float a, float b)
+{
+    return a*b;
+}
+
+
+float subraction(float a, float b)
+{
+    return a-b;
+}
+
+
