+++ calculator/functionalities.h
@@ -1,0 +1,14 @@
+#ifndef FUNCTIONALITIES_H
+#define FUNCTIONALITIES_H
+
+float addition(float a,float b);
+
+float division(float a, float b);
+
+float multiplication(float a,float b);
+
+float subraction(float a, float b);
+
+
+
+#endif // FUNCTIONALITIES_H
