--- calculator/multiplication/BUILD
@@ -1,9 +1,0 @@
-load("@rules_cc//cc:defs.bzl","cc_binary","cc_library")
-
-cc_library(
-
-    name = "multiplication",
-    srcs= ["multiplication.cpp"],
-    hdrs = ["multiplication.h"],
-    visibility = ["//visibility:public"],
-)
