--- calculator/multiplication/multiplication.cpp
@@ -1,7 +1,0 @@
-#include "multiplication.h"
-
-float multiplication(float a, float b)
-{
-
-    return a*b;
-}
