--- calculator/multiplication/multiplication.h
@@ -1,6 +1,0 @@
-#ifndef CALCULATOR_MULTIPLICATION_H
-#define CALCULATOR_MULTIPLICATION_H
-
-float multiplication(float a,float b);
-
-#endif // MULTIPLICATION_H
