--- calculator/subraction/BUILD
@@ -1,9 +1,0 @@
-load("@rules_cc//cc:defs.bzl","cc_binary","cc_library")
-
-cc_library(
-
-    name = "subraction",
-    srcs= ["subraction.cpp"],
-    hdrs = ["subraction.h"],
-    visibility = ["//visibility:public"],
-)
