--- calculator/subraction/subraction.cpp
@@ -1,6 +1,0 @@
-#include "subraction.h"
-
-float subraction(float a, float b)
-{
-    return a-b;
-}
