--- calculator/subraction/subraction.h
@@ -1,6 +1,0 @@
-#ifndef CALCULATOR_SUBRACTION_H
-#define CALCULATOR_SUBRACTION_H
-
-float subraction(float a, float b);
-
-#endif // SUBRACTION_H
